Refactore TimeoutIT to be a SimpleSuite test
Modified TimeoutIT to break the tests into separate
methods so that each test had its own timeout. The
three tests in a single method were taking just over
the timeout value.
